рефакторит LoginPage
